finished the first half of the IFML diagram classes

+++ Lib/src/main/java/FactoryElements/Lines/AbstractLine.java
@@ -1,20 +1,22 @@
 package FactoryElements.Lines;
 
+import FactoryElements.Interfaces.Line;
+
 /**
  * @author David Lindeman
  * Abstract line class with basic line functionality
  */
 
-public abstract class AbstractLine implements Line{
-    int leftXCord;
-    int leftYCord;
-    int rightXCord;
-    int rightYCord;
+public abstract class AbstractLine implements Line {
+    double leftXCord;
+    double leftYCord;
+    double rightXCord;
+    double rightYCord;
     String lineStyle;
 
 //    String color; //Use hex or RGB to represent?
 
-    public AbstractLine(int lX, int lY, int rX, int rY, String lStyle){
+    public AbstractLine(double lX, double lY, double rX, double rY, String lStyle){
         leftXCord = lX;
         leftYCord = lY;
         rightXCord = rX;
